+++ apps/api/src/app.module.ts
@@ -13,7 +13,11 @@ import { CashflowModule } from './cashflow/cashflow.module';
 import { CategoriesModule } from './categories/categories.module';
 import { CommonModule } from './common/common.module';
 import { DebtModule } from './debt/debt.module';
+import { DependentsModule } from './dependents/dependents.module';
+import { EstateDocumentsModule } from './estate-documents/estate-documents.module';
 import { ImportModule } from './import/import.module';
+import { IncomeSourcesModule } from './income-sources/income-sources.module';
+import { PlannedExpensesModule } from './planned-expenses/planned-expenses.module';
 import { InsuranceModule } from './insurance/insurance.module';
 import { NotificationsModule } from './notifications/notifications.module';
 import { PrismaModule } from './prisma/prisma.module';
@@ -47,6 +51,8 @@ import { TransactionsModule } from './transactions/transactions.module';
     CashflowModule,
     CategoriesModule,
     DebtModule,
+    DependentsModule,
+    EstateDocumentsModule,
     RecurringModule,
     ReadinessModule,
     RecommendationsModule,
@@ -54,6 +60,8 @@ import { TransactionsModule } from './transactions/transactions.module';
     SettingsModule,
     TransactionsModule,
     ImportModule,
+    IncomeSourcesModule,
+    PlannedExpensesModule,
     InsuranceModule,
   ],
   providers: [